内部仕様的にPKをIntで扱っていたので、実情に合わせて修正。 refs #29, #28

+++ app/src/main/java/net/formula97/andorid/car_kei_bo/data/LubMasterDao.kt
@@ -6,7 +6,7 @@ import android.arch.persistence.room.Query
 @Dao
 interface LubMasterDao : BaseDao<LubMaster> {
     @Query("SELECT * FROM LUB_MASTER WHERE RECORD_ID = :id")
-    fun findById(id: Long) : LubMaster?
+    fun findById(id: Int) : LubMaster?
 
     @Query("SELECT * FROM LUB_MASTER")
     fun findAll() : List<LubMaster>
